+++ packages/insomnia-api/src/index.ts
@@ -4,5 +4,6 @@ export * from './enterprise';
 export * from './trial';
 export * from './project';
 export * from './invite';
+export * from './organizations';
 
 export { configureFetch, type FetchConfig, ResponseFailError, isApiError } from './fetch';
