+++ packages/insomnia/src/routes/organization.$organizationId.permissions.tsx
@@ -1,12 +1,13 @@
+import { type FeatureList, getOrganizationFeatures, type Organization } from 'insomnia-api';
 import { href, redirect, type ShouldRevalidateFunctionArgs } from 'react-router';
 
 import { userSession } from '~/models';
-import { isScratchpadOrganizationId, type Organization } from '~/models/organization';
+import { isScratchpadOrganizationId } from '~/models/organization';
 import { insomniaFetch } from '~/ui/insomnia-fetch';
 import { createFetcherLoadHook } from '~/utils/router';
 
 import type { Route } from './+types/organization.$organizationId.permissions';
-import type { Billing, FeatureList } from './organization';
+import type { Billing } from './organization';
 
 export const fallbackFeatures = Object.freeze<FeatureList>({
   bulkImport: { enabled: false, reason: 'Insomnia API unreachable' },
@@ -44,11 +45,7 @@ export async function clientLoader({ params }: Route.ClientLoaderArgs) {
   }
 
   try {
-    const featuresResponse = insomniaFetch<{ features: FeatureList; billing: Billing } | undefined>({
-      method: 'GET',
-      path: `/v1/organizations/${organizationId}/features`,
-      sessionId,
-    });
+    const featuresResponse = getOrganizationFeatures({ organizationId, sessionId });
 
     return {
       featuresPromise: featuresResponse.then(res => res?.features || fallbackFeatures),
